Update VirusTotalReporter.py
This PR extends #115, to catch issues where a 200 is received but the report generation fails and presents the full error if the processor fails to decode the API's JSON response.

It also adds a new key: `VIRUSTOTAL_URL_ANALYSIS_ONLY`

If a value is passed to: `VIRUSTOTAL_URL_ANALYSIS_ONLY`, then only URL analysis will be performed, skipping the file analysis step completely.

This new key seems to be the only way we can get consistent anaylysis of the UnityHub media.

Co-Authored-By: Paul Cossey <paul_cossey@hotmail.com>